requirements.txt
--index-url https://download.pytorch.org/whl/cu128
--extra-index-url https://pypi.org/simple
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
# Core dependencies
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.0

# Unsloth (install from source or specific version)
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# Progress and CLI
rich>=13.7.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.1

# Evaluation metrics
evaluate>=0.4.1
rouge-score>=0.1.2
sacrebleu>=2.3.1
nltk>=3.8.1

# Logging and monitoring
tensorboard>=2.15.0
wandb>=0.16.0  # Optional, for W&B integration

# Utilities
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Optional: For better performance
